Move class definition of DebugInfoWriter to TraceUtil as well
[ghstack-poisoned]
