+++ torch/csrc/distributed/c10d/ProcessGroupNCCL.hpp
@@ -22,6 +22,7 @@
 #include <c10/cuda/CUDAGuard.h>
 #include <c10/cuda/CUDAStream.h>
 
+#include <torch/csrc/distributed/c10d/TraceUtils.h>
 #include <torch/custom_class.h>
 
 namespace c10d {
